tidy up server

+++ index.js
@@ -2,11 +2,11 @@ const fs = require("fs-extra");
 const path = require("path");
 
 const configFilePath = path.resolve(__dirname, ".env");
-if (!fs.existsSync(configFilePath)) {
+if (process.env.NODE_ENV !== "production" && !fs.existsSync(configFilePath)) {
   process.stdout.write("###### Creating .env file for first startup ########\n");
   fs.writeFileSync(configFilePath, fs.readFileSync(path.resolve(__dirname, ".env_bak")));
 }
 
 require("dotenv").config();
 
-require("./lib");
+require("./lib/server");
